Optimized ReadMe 4 C# Noobs
Reformed the readme in order to allow those whom did not understand what I thought was clearly obvious about how the koans were to be started and how they work. For those whom DID NOT READ THE README, this is another readme for you peeps. The Moradi is disappointed in the human race @ da moment....